+++ test/hihapi.test/UnitTests/Finance/FinanceDocumentLoanTest.cs
@@ -70,7 +70,7 @@ namespace hihapi.test.UnitTests
             {
                 fixture.InitHome5TestData(context);
             }
-            var account = context.FinanceAccount.Where(p => p.HomeID == hid && p.Status != FinanceAccountStatus.Closed).FirstOrDefault();
+            var account = context.FinanceAccount.Where(p => p.HomeID == hid && p.Status != (Byte)FinanceAccountStatus.Closed).FirstOrDefault();
             var cc = context.FinanceControlCenter.Where(p => p.HomeID == hid).FirstOrDefault();
 
             // 1. Create first Loan docs.
@@ -112,7 +112,7 @@ namespace hihapi.test.UnitTests
                     ? FinanceAccountCategory.AccountCategory_BorrowFrom
                     : FinanceAccountCategory.AccountCategory_LendTo,
                 Owner = user,
-                Status = FinanceAccountStatus.Normal,
+                Status = (Byte)FinanceAccountStatus.Normal,
             };
             var startdate = new DateTime(2020, 1, 10);
             var enddate = new DateTime(2021, 1, 10);
