+++ test/hihapi.test/UnitTests/Finance/FinanceTmpDPDocumentsControllerTest.cs
@@ -74,7 +74,7 @@ namespace hihapi.test.UnitTests
             {
                 fixture.InitHome5TestData(context);
             }
-            var account = context.FinanceAccount.Where(p => p.HomeID == hid && p.Status != FinanceAccountStatus.Closed).FirstOrDefault();
+            var account = context.FinanceAccount.Where(p => p.HomeID == hid && p.Status != (Byte)FinanceAccountStatus.Closed).FirstOrDefault();
             var cc = context.FinanceControlCenter.Where(p => p.HomeID == hid).FirstOrDefault();
 
             // 1. Create DP docs.
@@ -116,7 +116,7 @@ namespace hihapi.test.UnitTests
                         ? FinanceAccountCategory.AccountCategory_AdvancePayment
                         : FinanceAccountCategory.AccountCategory_AdvanceReceive,
                 Owner = user,
-                Status = FinanceAccountStatus.Normal,
+                Status = (Byte)FinanceAccountStatus.Normal,
             };
             var startdate = DateTime.Today;
             var enddate = DateTime.Today.AddMonths(6);
@@ -199,7 +199,7 @@ namespace hihapi.test.UnitTests
 
             // Check the account status
             var account2 = context.FinanceAccount.Where(p => p.HomeID == hid && p.ID == dpacntid).FirstOrDefault();
-            Assert.True(account2.Status == FinanceAccountStatus.Closed);
+            Assert.True(account2.Status == (Byte)FinanceAccountStatus.Closed);
 
             CleanupCreatedEntries();
